move tool scope behind ToolView and conflicts onto bindings
`ToolBindings` becomes `ToolSet` behind `Arc<Mutex<ToolSet>>`, with the read-only `ToolView` trait implemented directly on the mutex, and `RunContext` holds `Arc<dyn ToolView>` created empty at run start, so the tool set rides across the H1-to-H2 boundary with no fork. Near-duplicate conflicts are computed per `tools.bind` and recorded symmetrically on both bindings as `conflicts: Vec<Conflict>`.

- Scope validation becomes a local check: a clash errors when both halves of a recorded conflict enter one model-visible scope, and `ToolAnalysis` is deleted.
- `ToolResolver` gains a `near_duplicates` method whose default reports no pairs, so a resolver without similarity knowledge records no conflicts.
- `Conflict` compares the similarity score by bits, keeping equality reflexive at NaN for `ToolBinding`'s `Eq`.
- The never-emitted `ToolRegistryValidationStarted`, `ToolRegistryValidationSucceeded`, and `ToolRegistryValidationFailed` observation variants are removed.
- A new test pins that two clashing binds each record the other alias with the picker's score.

+++ crates/promptforge-core/src/tools/registry.rs
@@ -126,16 +126,16 @@ impl ToolCatalog {
 #[derive(Debug)]
 #[non_exhaustive]
 pub(crate) struct NearDuplicateDiagnostic {
-    /// The first prompt-local alias in picker catalog pair order.
+    /// The first prompt-local alias in scope order.
     pub(crate) first_alias: String,
     /// The first stable identity.
     pub(crate) first_id: ToolId,
-    /// The second prompt-local alias in picker catalog pair order.
+    /// The second prompt-local alias in scope order.
     pub(crate) second_alias: String,
     /// The second stable identity.
     pub(crate) second_id: ToolId,
-    /// The cosine similarity reported by the picker.
-    pub(crate) similarity: f32,
+    /// The cosine similarity the picker reported at bind time.
+    pub(crate) similarity: f64,
 }
 /// A stable, matchable classification of a [`ToolCatalogError`].
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
